add MSFT_texture_dds extension support
This logic is duplicated from WEBP's logic, as we only check for
a new mime type.
Checks the MSFT_texture_dds is enabled, and that the mime type is
matching.
For the DDS decoder, only checks the magic, no decoding available.

This allows GLTF with this extension to pass validation.